+++ skedge-frontend/src/components/HeaderPage.js
@@ -8,7 +8,8 @@ import {withRouter} from 'react-router-dom';
 class HeaderPage extends Component {
   constructor(props) {
     super(props);
-
+    this.state = { authToken : props.location.authToken };
+    
 
     this.handleRecordButtonClick = this.handleRecordButtonClick.bind(this);
     this.handleLogout = this.handleLogout.bind(this);
@@ -17,7 +18,11 @@ class HeaderPage extends Component {
 
   handleRecordButtonClick(){
     console.log("Record");
-    this.props.history.push('/record')
+    this.props.history.push({
+      pathname: '/record',
+      authToken: this.state.authToken
+    }); 
+
   }
 
   handleLogout(){
